pyproject.toml
[tool.poetry]
name = "09-04-2024"
version = "0.1.0"
description = ""
authors = ["miraclemenikelechi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
pydantic = {extras = ["email"], version = "^2.6.4"}
pydantic-extra-types = "^2.6.0"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"
phonenumbers = "^8.13.34"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
